Refactor complex asin/acos/asinh/acosh to use asin_acos_kernel. Improve real acos accuracy. (#2496)
As in the title.

This PR introduces asin_acos_kernel operation that implements modified
Hull et al algorithm for evaluating complex asin, acos, asinh, and acosh
operations. This task corresponds to the refactor request in
https://github.com/openxla/stablehlo/pull/2411#discussion_r1662810017
and it resolves
https://github.com/pearu/functional_algorithms/issues/15.

In addition, the PR also improves the real acos accuracy as a fix to
https://github.com/openxla/stablehlo/issues/2452 . As a result, the
accuracy of single-precision real acos improved as follows (using 1
million samples):
```
                             main       this PR
ULP difference == 0 count is 971713  -> 999045
ULP difference == 1 count is 28158   ->    962
ULP difference == 2 count is 54      ->      0
ULP difference == 3 count is 20      ->      0
ULP difference >= 4 count is 62      ->      0
```


The cause of the revert by PR
https://github.com/openxla/stablehlo/pull/2449 (`acos(-1) -> 0` while
expecting `pi`) is fixed in functional_algorithms which now also
generates extra samples that correspond to limiting values of real acos.
For instance, `acos(-1)->pi`, `acos(nextafter(-1, -inf))->nan`, and
`acos(nextafter(-1, int)-> approx pi` are now included in the tests.


The required functional_algorithms version is now 0.9 which includes a
fix to `fa.utils.real_samples` to return samples that are distributed
uniformly with respect to ULP differences of neighboring samples. This
fix lead to updates of `stablehlo/tests/math/*.mlir`.