Move exported files behind filegroups (#2504)
It is problematic to include StableHLO source files from non-StableHLO
projects. It makes renaming files and dependencies much more difficult.

Hiding file names behind a target that lives in StableHLO should help
with the file renames case. There isn't a good way to keep a dependency
list in StableHLO, but the only dependency of python bindings should be
CAPI so that shouldn't be too problematic.