Workflow to run tutorials in CI (#2663)
This PR creates a workflow to execute the Jupyter notebooks in
`docs/tutorials`. Couple of notes:
- It installs dependencies as defined in the notebook with `!pip install
...`, making it a little slow. We can instead create a
`requirements.txt` and [cache
dependencies](https://github.com/actions/setup-python#caching-packages-dependencies)
for efficient runs, but then we won't be testing the actual user
workflows running the tutorials on kaggle/colab
- The workflow only _executes_ the notebooks to make sure there are no
errors. The notebooks are not altered. We can also update the notebooks
in-place with the freshly executed versions for consistent output. The
only concern here is the workflow might fail to push commits to the PR
if it were opened from forks due to permission issues.

The PR also updates `savedmodel-embed.ipynb`; `target_version` is now a
required argument in `stablehlo_to_tf_saved_model()`

Tested on my fork:
https://github.com/pavithraes/stablehlo/actions/runs/12239259032/job/34139262358#step:5:1

---------

Signed-off-by: Pavithra Eswaramoorthy <pavithraes@outlook.com>