Add source location downgrade to compatibility expander (#2725)
Fixes https://github.com/openxla/stablehlo/issues/2724

Compatibility expander is an opt-in pass used in PJRT/ODML/etc to try to
provide forward compatibility even when a new feature is used.

These can be costly or impact performance as they usually decompose a
single op into many, etc.

Note - This is only run when requesting to downgrade to a version before
1.8.5, meaning it will not impact any PJRT users today, only for jax2tf
or other export workflows.