Fix- PR "cmake-build" Check stuck at "pending"  Status Issue (#2774)
[Draft]
I don't know why this fix is working!

workflow name: `CMake Build`
JobID: `cmake-build`

Issue: As soon as Checks start running for a PR, `cmake-build` check is
going to a pending status. Never changed to `in-progress` or
`completed`. Note - There is a already a check with `CMake Build /
cmake-build (pull_request)` which runs alright.

In Branch protection setting of the repo, `cmake-build` is marked as
`required`. If we mark it as not `required`, this extra `cmake-build`
disappears.

Why this fix: updating....