Changes to allow StableHLO to be added as a Bzlmod dep (#2743)
To add StableHLO as a dependency in the WORKSPACE (due to it depending
on LLVM, which doesn't support Bzlmod yet) but in a Bazel project that
uses Bzlmod then a patch is required to get it to work.

The WORKSPACE would have something like this in it:
```
git_repository(
    name = "stablehlo",
    commit = "48a1e14edc8219577fcad53de1924876f855f431",
    patch_args = ["-p1"],
    patches = [
        "//patches:stablehlo/00-bazel-dep.patch",
    ],
    remote = "https://github.com/openxla/stablehlo.git",
)
```

This PR is the contents of that patch file.