Add a pattern for folding `stablehlo.while` no-ops
If a `stablehlo.while` op's condition is a constant with a value of
`false`, replace the op with its input.

Make an exception for preserving dead ops with side effects, which can't
currently be DCE'd safely despite being unreachable. This is because
dependent code is currently allowed to depend on the existence of ops
with hypothetical side effects that show up in the unoptimized IR but
can't actually occur.