Optionally DCE loops more liberally (opt-in).
Firstly, this fixes a bug where we assumed an `has_side_effect`
attribute meant that a `custom_call` had every side effect even though
the attribute was documented as semantically `false` if unset.

Secondly, added an option (disabled by default) to assume, for purposes
of eliminating dead `while` ops, that ops are pure unless they say
otherwise via the MLIR `MemoryEffects` interface. The default behavior
is still to play it safe and assume ops have side effects if they don't
specify one way or the other, but enabling the new option can result in
significantly smaller IR. (Note that MHLO optimizations already make
this assumption when removing dead `while` ops, which suggests that it's
typically safe to enable the option.)