Fold `case` ops with constant branch indices (#2843)
When a `case` op is given a constant branch-selection index, replace the
`case` op with the contents of the always-active branch. If any branches
are non-trivially dead (i.e. dead but not functionally pure), the `case`
op will not be deleted but will always execute a now-empty branch (since
the original contents of the active branch have been inlined).