Optimize folding transposes that operate on splats
Instead of expanding and iterating over the splat to create the new
constant, we now just replace the splat constant's dimensions as
specified by the transpose op.

When tested on the same input program that brought this issue to light,
this fix improved the optimizer's execution time from 8.68 s to 0.80 s,
a 985% speedup.