Fix Lowering stablehlo.dynamic_broadcast_in_dim to linalg mlir fails with dynamic shapes (#2833)
If the input and output have a dynamic shape such that the rank of input
matches the rank of output and shapes are compatible (i.e. no broadcasting is
needed), then the dynamic broadcast in dim is effectively a copying input to
output In such cases, we can lower it to a simple linalg.generic operation.
It can later be canonicalised and bufferised accordingly.