Add parentheses to silence `a || b && c` warning
The CMake build was failing due to `-Werror,-Wlogical-op-parentheses`
since the simplification patterns included an expression of the form
`a || b && c`, which the compiler decided was unclear since it relied on
understanding the operator precedence of `&&` vs. `||`. This resolves
the warning by adding parentheses, e.g. `a || (b && c)`.